refactor(mobile): add new NetworkDisplay row
