+++ apps/mobile/src/features/HistoryTransactionDetails/components/history-views/HistoryGenericView.tsx
@@ -1,21 +1,18 @@
 import React from 'react'
 import { Container } from '@/src/components/Container'
 import { View, YStack, Text, H3 } from 'tamagui'
-import { Logo } from '@/src/components/Logo'
 import { TransactionHeader } from '@/src/features/ConfirmTx/components/TransactionHeader'
 import {
   TransactionData,
   Transaction,
   TransactionDetails,
 } from '@safe-global/store/gateway/AUTO_GENERATED/transactions'
-import { useAppSelector } from '@/src/store/hooks'
-import { selectChainById } from '@/src/store/chains'
-import { RootState } from '@/src/store'
-import { useDefinedActiveSafe } from '@/src/store/hooks/activeSafe'
+
 import { Address } from '@/src/types/address'
 import { HistoryAdvancedDetailsButton } from '@/src/features/HistoryTransactionDetails/components/HistoryAdvancedDetailsButton'
 import { useTransactionType } from '@/src/hooks/useTransactionType'
 import { HashDisplay } from '@/src/components/HashDisplay'
+import { NetworkDisplay } from '../shared'
 
 interface HistoryGenericViewProps {
   txId: string
@@ -25,9 +22,6 @@ interface HistoryGenericViewProps {
 }
 
 export function HistoryGenericView({ txId, txInfo, txData, executedAt }: HistoryGenericViewProps) {
-  const activeSafe = useDefinedActiveSafe()
-  const activeChain = useAppSelector((state: RootState) => selectChainById(state, activeSafe.chainId))
-
   const recipientAddress = txData?.to?.value as Address
 
   // Create a transaction object for the hook - need full Transaction type
@@ -64,14 +58,7 @@ export function HistoryGenericView({ txId, txInfo, txData, executedAt }: History
               </View>
             )}
 
-            <View alignItems="center" flexDirection="row" justifyContent="space-between">
-              <Text color="$textSecondaryLight">Network</Text>
-
-              <View flexDirection="row" alignItems="center" gap="$2">
-                <Logo logoUri={activeChain?.chainLogoUri} size="$6" />
-                <Text fontSize="$4">{activeChain?.chainName}</Text>
-              </View>
-            </View>
+            <NetworkDisplay />
 
             <HistoryAdvancedDetailsButton txId={txId} />
           </Container>
