+++ apps/mobile/src/features/HistoryTransactionDetails/components/history-views/HistoryRemoveSigner.tsx
@@ -1,15 +1,12 @@
 import React from 'react'
 import { Container } from '@/src/components/Container'
 import { View, YStack, Text, H3 } from 'tamagui'
-import { Logo } from '@/src/components/Logo'
 import { TransactionHeader } from '@/src/features/ConfirmTx/components/TransactionHeader'
 import { MultisigExecutionDetails } from '@safe-global/store/gateway/AUTO_GENERATED/transactions'
 import { NormalizedSettingsChangeTransaction } from '@/src/features/ConfirmTx/components/ConfirmationView/types'
 import { HistoryAdvancedDetailsButton } from '@/src/features/HistoryTransactionDetails/components/HistoryAdvancedDetailsButton'
-import { useAppSelector } from '@/src/store/hooks'
-import { selectActiveChain } from '@/src/store/chains'
 import { HashDisplay } from '@/src/components/HashDisplay'
-import { ThresholdChangeDisplay } from '../shared/ThresholdChangeDisplay'
+import { ThresholdChangeDisplay, NetworkDisplay } from '../shared'
 
 interface HistoryRemoveSignerProps {
   txId: string
@@ -19,8 +16,6 @@ interface HistoryRemoveSignerProps {
 }
 
 export function HistoryRemoveSigner({ txId, txInfo, executionInfo, executedAt }: HistoryRemoveSignerProps) {
-  const activeChain = useAppSelector(selectActiveChain)
-
   return (
     <>
       <TransactionHeader
@@ -44,13 +39,7 @@ export function HistoryRemoveSigner({ txId, txInfo, executionInfo, executedAt }:
 
             <ThresholdChangeDisplay txInfo={txInfo} executionInfo={executionInfo} />
 
-            <View alignItems="center" flexDirection="row" justifyContent="space-between">
-              <Text color="$textSecondaryLight">Network</Text>
-              <View flexDirection="row" alignItems="center" gap="$2">
-                <Logo logoUri={activeChain?.chainLogoUri} size="$6" />
-                <Text fontSize="$4">{activeChain?.chainName}</Text>
-              </View>
-            </View>
+            <NetworkDisplay />
 
             <HistoryAdvancedDetailsButton txId={txId} />
           </Container>
