+++ apps/mobile/src/features/HistoryTransactionDetails/components/history-views/HistoryTokenTransfer.tsx
@@ -1,17 +1,15 @@
 import React from 'react'
 import { Container } from '@/src/components/Container'
 import { View, YStack, Text, H3 } from 'tamagui'
-import { Logo } from '@/src/components/Logo'
 import { TransactionHeader } from '@/src/features/ConfirmTx/components/TransactionHeader'
 import { TransferTransactionInfo } from '@safe-global/store/gateway/AUTO_GENERATED/transactions'
 import { useTokenDetails } from '@/src/hooks/useTokenDetails'
-import { useAppSelector } from '@/src/store/hooks'
-import { selectActiveChain } from '@/src/store/chains'
 import { Address } from '@/src/types/address'
 import { TokenAmount } from '@/src/components/TokenAmount'
 
 import { HistoryAdvancedDetailsButton } from '@/src/features/HistoryTransactionDetails/components/HistoryAdvancedDetailsButton'
 import { HashDisplay } from '@/src/components/HashDisplay'
+import { NetworkDisplay } from '../shared'
 
 interface HistoryTokenTransferProps {
   txId: string
@@ -20,7 +18,6 @@ interface HistoryTokenTransferProps {
 }
 
 export function HistoryTokenTransfer({ txId, txInfo, executedAt }: HistoryTokenTransferProps) {
-  const activeChain = useAppSelector(selectActiveChain)
   const { value, tokenSymbol, logoUri, decimals } = useTokenDetails(txInfo)
 
   const isOutgoing = txInfo.direction === 'OUTGOING'
@@ -63,14 +60,7 @@ export function HistoryTokenTransfer({ txId, txInfo, executedAt }: HistoryTokenT
               <HashDisplay value={address as Address} />
             </View>
 
-            <View alignItems="center" flexDirection="row" justifyContent="space-between">
-              <Text color="$textSecondaryLight">Network</Text>
-
-              <View flexDirection="row" alignItems="center" gap="$2">
-                <Logo logoUri={activeChain?.chainLogoUri} size="$6" />
-                <Text fontSize="$4">{activeChain?.chainName}</Text>
-              </View>
-            </View>
+            <NetworkDisplay />
 
             {isOutgoing && <HistoryAdvancedDetailsButton txId={txId} />}
           </Container>
