+++ apps/mobile/src/features/HistoryTransactionDetails/components/shared/index.ts
@@ -1,1 +1,2 @@
 export { ThresholdChangeDisplay } from './ThresholdChangeDisplay'
+export { NetworkDisplay } from './NetworkDisplay'
